refactor(cli): simplify evaluation module names
- Rename evaluation flow modules to plain responsibility-based names

  (agent-evaluation, file-evaluation, result-routing, issue-output)

- Update orchestrator and internal imports to match the new

  naming convention without changing runtime behavior
